Mostly wire up core class method binding with IndyMethodFactory.
Core class methods now can be bound using MethodHandles
exclusively. This continues work to allow us to strip out
generated invokers and populators and ship a smaller jar for Java
7+ with invokedynamic available.

TODO:

* No arity-checking logic for variable-arity paths
* No arity error for variable-arity call with no good target
* Trace calls

Performance with populators and invokers removed and this factory
active is 2-2.5x slower than generated populators and invokers,
but that could be an artifact of the extra work being done by
the default populators. We may need to look into a way to do more
work at build time to determine the eventual structure of MH-based
invokers, rather than programmatically doing it on every boot.