Fix embedding test failures introduced by Callback removal.
A single rogue call to Ruby.callbackFactory I removed from
RubySystemCallError was to blame for this. callbackFactory forced
the JRubyClassLoader stored in Ruby to be initialized, and without
that the field was left null. External uses all called via
Ruby.getJRubyClassLoader, which lazily initialized it...but
somethign in embedding or tests for it called through
compileAndLoadFile, which accesses it directly. Because the
JRubyClassLoader is almost always needed to run JRuby, I have
added a call to Ruby.getJRubyClassLoader into the Ruby constructor
to ensure it has been initialized. Tests pass now.