Fix recent OpenSSL failures by updating CA certs used to verify.
It seems that www.amazon.com changed the CA cert it uses for its
ssl connection, and that caused us to start failing tests. After
some investigation, I figured out which cert we needed and what
the filename needed to be in the ca_path test. I have added a
README.txt to the ca_path and comments to all the related certs.
I also removed a cert file that appeared to be duplicate and added
a more explicitly incorrect cert for the test that is supposed to
fail verification.