Fix JRUBY-7145 by checking for drive letter in Dir.mkdir.
Dir.mkdir did not have any checks for the case where the path
passed in was a drive letter, as happens when FileUtils.mkdir_p
breaks up the path into components. It should obviously not
attempt to make a directory for a drive letter, so I have made
this a no-op.