Go back to known-working-but-slower Exchanger impl of #next.
The exchanger impl is still slower than the pure-Ruby +
queue impl, but it does not leak threads and appears to
reliably shut down abandoned Enumerators. We can iterate on
improving perf of this impl (and applying that knowledge to
Fiber perf) post 1.7.4.

This reverts commits d3cf2954b38110da9f8d63f46c0757ad9f451c8d
and f19a8376453656c113b9ef9adf4b365eabaf37d6.