Remove bad assert
Seems this assert was typoed, and was accidentally asserting the
nullness of something which is definitely not null.  This was causing
a number of false negatives when the tests were run with assertions
enabled.

Note that this commit completely removes the null checking since it
can't do any good: this code would NPE earlier if line was null.