Rework frame management, and backref/lastline along with it.
These changes should bring us more in line with the way other
implementations manage "frame" data and bindings that capture it.
It should also reduce the cost of framing, binding captures,
blocks and procs, and operations that manipulate blocks and procs
like define_method and instance_eval.

The new logic works as follows:

* Rather than a binding immediately duplicating a frame, it
  instead grabs a reference to it and marks it as captured.
* When popping a frame, if it has been marked captured it is
  discarded rather than being cleared. This moves the cost of
  creating a binding to the pop of the surrounding method, which
  reduces the cost of nested bindings re-cloning the frame many
  times.
* Bindings that are still active (have captured a frame that has
  not been popped) will update the method's frame data in-place.
* Once the frame has been popped, it will live only in that
  binding.

This commit passes mspec ci and test:mri19, but I am continuing
to test it for a while.

Note that this commit also moves backref and lastline to
the frame. They have always been frame-local, but for a long time
we kept them on the scope so we could avoid touching the frame for
methods that did not need it. However, because we used the scope,
it was difficult or impossible to isolate backref and lastline
changes to the block. Moving to the frame, which has a lifespan
only as long as the method that activated it, makes isolation
easier.

This should also make it easier to fix thread-safety issues with
backref/lastline, as reported in JRUBY-7167.