Allow disabling 8-bit transcode error with :undef. Fixes #616.
ASCII-8BIT is basically treated as binary for general purposes, so
when MRI is asked to transcode ASCII-8BIT data to a text encoding,
it requires the data to conform to US-ASCII byte range or raises
an error. On MRI, this error is UndefinedConversionError, and the
user can opt to replace those high bytes by specifying :undef =>
:replace as a transcoding option.

On JRuby, we originally had a hard fail for this error that could
not be disabled. I moved the fail into the transcode logic by
treating ASCII-8BIT input encoding as US-ASCII for purposes of
text transcoding. This causes the Java transcoding subsystem to
raise an error.

Unfortunately, the error raised is MalformedInputException,
indicating that the input data is not well-formed character data.
I believe this is *correct* but it means we would have to enable
replacement using :invalid rather than :undef. To support compat
with MRI, I chose to set the :invalid action to the :undef action
in this situation, which allow :undef to replace high bytes as
required. I also filed http://bugs.ruby-lang.org/issues/8630
because I believe "invalid" is the proper treatment of high-byte
ASCII-8BIT characters when attempting to do a text transcode.

Fixes #616.