Change `--properties` output for null default opts [skip ci]
Fixes #949

For options whose default is `null`, it is confusing to print out the
string `null`, since it is not a likely value that the user can use.
While the message makes it clear that the value will be interpreted as a
String, it is a good idea to leave it blank instead, so that the user
can fill it themselves.