Never use core runtime cache for user classes. Fixes #1004.
The "index" of a RubyModule is only non-zero when it's one of our
known native core classes. All other classes end up with zero.
Since we use this index to offset into the core method cache (in
Ruby.getRuntimeCache() or ThreadContext.runtimeCache) we must
check for index == zero and not use the cache in that case (e.g.
in Helpers.invokedynamic() which calls Helpers.getMethodCached().