Make Thread#backtrace a bit more thread-safe.
The original NPE for #406 was due to the thread not having started
running yet; its "context" reference was null. I added logic to
check for that and not try to produce a backtrace.

In addition, I discovered that the building of the trace was also
not threadsafe, since the "backtrace" and "backtraceIndex" fields
in ThreadContext could be updated at the same time by the original
thread. My changes here should make it less likely that backtrace
building will walk off the end of the "backtrace" array, but there
will still be cases where the index and the array get out of sync
and the backtrace contains a couple bogus lines. It is unclear
to me whether we should forcibly prevent the target thread from
updating these fields while the backtrace-generating thread is
generating, since it would surely introduce overhead into the
normal backtrace updating process.

Fixes #406.