Conditionally boot the extension only if not booted already.
Rather than using the older method for booting the jopenssl
extension, we now directly load the library. However, this does
not enlist in require logic, so it can load twice if done
unconditionally. This commit modifies that loading to only occur
if the jopenssl.jar library has not been loaded before.

Fixes #1071