Turn off ObjectProxyCache for 9k.
ObjectProxyCache names a weak map used by JRuby to attempt to use
the same wrapper for a given Java object every time it enters Ruby
space. This was also intended to guarantee that the instance vars
and singleton class associated with a Java object would be
preserved even if the Java object left Ruby and came back.

In practice, the cache added significant overhead to all callouts
to Java, since each call returning a non-coercible object must
trigger a proxy wrapper search. It also did not guarantee that
ivars and singletons would be preserved if the Java object left
Ruby space, since the cache must be weak. It only guaranteed that
the same Java object re-entering Ruby without any chance for the
proxy wrapper to go away would get the same proxy wrapper.

In JRuby 1.7 we began to deprecate the ObjectProxyCache, adding
a mechanism to opt-in rather than imposing that overhead on all
Java types called from Ruby. This commit proceeds from that
deprecation to disable ObjectProxyCache by default.