Remove old Generator impl, move toward unifying coroutine logic.
The old Enumerator#next logic used our internal Generator impl,
which proved eventually to be leaky and slow. As of JRuby 1.7.5,
Enumerator#next moved to a Java implementation that was both
faster and better about cleaning up dangling instances, but the
block form of Enumerator.new still used the old Generator logic.

This commit eliminates the old coroutine-based Generator in favor
of a trivial version matching Ruby 1.9+. The block forms of
Enumerator.new now create this trivial Generator and then fall
into the same "Nexter" logic introduced in 1.7.5 for #next.